+++ CHANGELOG.md
@@ -4,6 +4,13 @@ This file is a running track of new features and fixes to each version of the pa
 
 This project follows [Semantic Versioning](http://semver.org) guidelines.
 
+## v3.6.1-beta
+
+### Fixed
+
+- IP Addresses being clipped if too long in the UI
+- Awkward delete address modal
+
 ## v3.6.0-beta
 
 ### Fixed